Authentication sections added
