+++ package.json
@@ -13,12 +13,14 @@
     "next": "10.0.4",
     "postcss": "^8.2.2",
     "react": "17.0.1",
+    "react-custom-scroll": "^4.2.1",
     "react-dom": "17.0.1",
     "tailwindcss": "^2.0.2"
   },
   "devDependencies": {
     "@types/node": "^14.14.19",
     "@types/react": "^17.0.0",
+    "@types/react-custom-scroll": "^4.2.0",
     "typescript": "^4.1.3"
   }
 }
