+++ src/pages/index.tsx
@@ -1,6 +1,3 @@
-import Head from "next/head"
-import styles from "../styles/Home.module.css"
-
 export default function Home() {
-  return <div>Dashboard</div>
+  return <div>Dashboard</div>;
 }
