+++ src/pages/students/personal/residential-address/index.tsx
@@ -1,5 +1,7 @@
-import React from "react";
-import FormFooter from "../../../components/students/FormFooter";
+import React from 'react';
+import FormFooter from '../../../../components/students/FormFooter';
+import StudentHeader from '../../../../components/students/StudentHeader';
+import StudentsLayout from '../../../../layouts/StudentsLayout';
 
 export const ResidentialForm = () => {
   return (
@@ -152,7 +154,7 @@ export const ResidentialForm = () => {
               name="city-town-suburb"
               id="city-town-suburb"
               className="shadow-sm focus:ring-indigo-500 focus:border-indigo-500 block  sm:text-sm border-gray-300 rounded-md"
-              style={{ width: "31rem" }}
+              style={{ width: '31rem' }}
             />
           </div>
         </div>
@@ -221,7 +223,7 @@ export const ResidentialForm = () => {
           id="country"
           name="country"
           className="mt-1 block pl-3 pr-10 py-2 text-base text-gray-900 border-gray-300 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md"
-          style={{ width: "31rem" }}
+          style={{ width: '31rem' }}
           placeholder="Select country"
         >
           <option>USA</option>
@@ -235,83 +237,24 @@ export const ResidentialForm = () => {
 
 const index = () => {
   return (
-    <div className="flex flex-col w-full overflow-y-scroll bg-bgPrim">
-      <div className="py-5 px-md">
-        <nav className="flex" aria-label="Breadcrumb">
-          <ol className="flex items-center space-x-2">
-            <li>
-              <div>
-                <a
-                  href="#"
-                  className="ml-0 text-sm font-medium text-blue-500 hover:text-blue-700"
-                >
-                  Home
-                </a>
-              </div>
-            </li>
-            <li>
-              <div className="flex items-center">
-                <svg
-                  width="5"
-                  height="12"
-                  viewBox="0 0 5 12"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M4.1804 0.863636H3.04261L0.230114 11.3125H1.3679L4.1804 0.863636Z"
-                    fill="#212121"
-                  />
-                </svg>
-                <a
-                  href="#"
-                  className="ml-2 text-sm font-medium text-gray-500 hover:text-gray-700"
-                >
-                  New Application
-                </a>
-              </div>
-            </li>
-            <li>
-              <div className="flex items-center">
-                <svg
-                  width="5"
-                  height="12"
-                  viewBox="0 0 5 12"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M4.1804 0.863636H3.04261L0.230114 11.3125H1.3679L4.1804 0.863636Z"
-                    fill="#212121"
-                  />
-                </svg>
-                <a
-                  href="#"
-                  aria-current="page"
-                  className="ml-2 text-sm font-medium text-gray-500 hover:text-gray-700"
-                >
-                  Personal
-                </a>
-              </div>
-            </li>
-          </ol>
-        </nav>
-        <p className="text-3xl font-bold leading-9 mt-3">Residential Address</p>
-        <p className="text-gray-500 text-sm font-medium">
-          Ultrices sagittis justo vitae, rhoncus habitant fermentum cursus
-          interdum sed.
-        </p>
-      </div>
-      <div className=" pb-0 px-md bg-bgPrim">
-        <div className="w-1056px px-4 py-8 bg-white rounded-md shadow-md">
-          <ResidentialForm />
+    <StudentsLayout>
+      <div className="flex flex-col w-screen overflow-y-scroll bg-bgPrim">
+        <StudentHeader
+          text="Residential Address"
+          label="Ultrices sagittis justo vitae, rhoncus habitant fermentum cursus
+        interdum sed."
+        />
+        <div className="py-sm px-md bg-bgPrim">
+          <div className="w-1056px px-4 py-8 bg-white rounded-md shadow-md">
+            <ResidentialForm />
+          </div>
         </div>
+        <FormFooter
+          forwardLink="/students/personal/postal-address"
+          backLink="/students/personal/visa-details"
+        />
       </div>
-      <FormFooter
-        forwardLink="/personal/postal-address"
-        backLink="/personal/visa-details"
-      />
-    </div>
+    </StudentsLayout>
   );
 };
